requirements.txt
opencv-python
tensorflow~=2.15.0
numpy
flask
pillow
